Switch pre-commit update workflow to use two tokens
In this merge, we go back to the simplified workflow that does
not use conda to install `gh` and instead use the latest `gh`
from github-actions.

This requires using a personal access token (PAT) to set reviewers
and labels, so one has been added to the repo.

We still use the github-actions bot to open the PR so it doesn't
get opened as @xylar (and @xylar can still review).